Clean up the MSVC build and fix compilation errors on older versions.
Versions prior to Visual Studio 2010 do not support _Interlocked*8/16()
which means they need to be disabled on the 64-bit build. They can be
enabled on the 32-bit build because we can just use inlined assembly.

With this commit, support has been improved for 32-bit builds using
versions of Visual Studio prior to 2010. 64-bit builds using the same
compiler will continue to throw an error if attempting to use 8- and
16-bit atomics. This is because the 64-bit MSVC compiler does not
support inlined assembly, without which we cannot implement a fallback
for when _Interlocked*8/16() are not implemented.