Big code restructure.
With this change, it's now practical to implement atomics with the use
of a global lock when a native instruction is unavailable. When this
global lock is used, c89atomic.c will need to be compiled and linked
into the program.

This change also makes it easier to test the various code paths without
having to switch compilers.